refactor: move otp into authentication
