+++ src/apps/backend/modules/authentication/internals/otp/internal/otp_util.py
@@ -2,9 +2,9 @@ import random
 import string
 from typing import Any
 
-from modules.otp.internal.store.otp_model import OTPModel
+from modules.authentication.internals.otp.internal.store.otp_model import OTPModel
 from modules.config.config_service import ConfigService
-from modules.otp.types import OTP
+from modules.authentication.internals.otp.types import OTP
 
 
 class OTPUtil:
