+++ tests/modules/access_token/base_test_access_token.py
@@ -3,7 +3,7 @@ from typing import Callable
 
 from modules.authentication.rest_api.authentication_rest_api_server import AuthenticationRestApiServer
 from modules.account.internal.store.account_repository import AccountRepository
-from modules.otp.internal.store.otp_repository import OTPRepository
+from modules.authentication.internals.otp.internal.store.otp_repository import OTPRepository
 
 
 class BaseTestAccessToken(unittest.TestCase):
