+++ tests/modules/access_token/test_access_token_api.py
@@ -8,8 +8,8 @@ from modules.account.types import (
     CreateAccountByUsernameAndPasswordParams,
     PhoneNumber,
 )
-from modules.otp.otp_service import OTPService
-from modules.otp.types import CreateOTPParams, OTPErrorCode, VerifyOTPParams
+from modules.authentication.internals.otp.otp_service import OTPService
+from modules.authentication.internals.otp.types import CreateOTPParams, OTPErrorCode, VerifyOTPParams
 from server import app
 from tests.modules.access_token.base_test_access_token import BaseTestAccessToken
